Refactor SQLite database generation
- Move SQL query constants, map of queries and map of table
  to map of query set  from `config` package to `dbqs` package

- Move SQLite preparation code out of main to `dbqs` package as
  new `PrepareSQLiteDB` func

- Update misc doc comments

I'm not entirely happy with the results, but it seems a bit
better than before. Once I refactor more of the codebase I will
likely return to this code and refine further.

In particular, the `config` package now depends directly on the
`dbqs` package for constants and types. This was done so that
the `dbqs` package would be aware of those now relocated types
and constants, but this dependency is questionable.