Update docs to cover contrib content, apps story
By "story", I refer to the real-world use case provided by
the `contrib` content and this updated documentation.

By providing a specific deployment scenario, the hope is that
this application's usage will be made clearer. Further
refinements to the directions are expected. This should be
considered a first draft submitted for review in order to
collect feedback.

refs GH-13
refs GH-17