Emit successful config load msg at startup
This can be useful for telling at a glance which configuration
file was detected and used. Because this behavior is automatic
and uses a hierarchy, a configuration file can be used that
the user might not expect.

This behavior is subject to change in the future. Emitting the
config file used is a useful workaround for now.