Add missing coverage for queries key and subkeys
The configuration documentation was missing coverage for
this config file structure. A new table is used in an
effort to make the coverage easier to read.

An example code block is provided to illustrate the structure.

Also, I removed the `Flag Name` column and duplicated the
description of the shared logging settings from the flags table.
This seemed a better use of horizontal space.

refs GH-13