Disable printing Go version provided by GHAWs VM
When reviewing the latest Actions results I saw that Go 1.14.9
was "used" by the "Build static binaries using Docker images"
job. This initially confused me until I realized that this was
emitted from within the GitHub Actions virtual machine
environment and not from within the Docker containers used
for the job.

This commit disables printing that information and notes why
in an effort to reduce confusion going forward.